requirements.txt
Django>=3.0,<4.0
psycopg2-binary>=2.8
celery>=5.0.0
djangorestframework==3.12.2
django-cors-headers==3.7.0
pika==1.1.0
asyncpg==0.22.0
fastapi==0.63.0
uvicorn==0.13.4
strawberry-graphql
requests
http3
redis
pymongo
pytest
websockets
django-ninja
orjson
markdown
django-filter
pyyaml
django-environ
kafka-python